Remove __pmLock_libpcp locks from around calls to __pmGetAddrInfo().
One of the benefits of using getaddrinfo() and getnameinfo(),
as opposed to their deprecated counterparts, gethostbyname() and
gethostbyaddr(), is that the new APIs are re-entrant. As such, it
is not necessary to enforce locks around calls to these functions
nor the iteration over the returned data.

This commit removes the remaining locks around __pmGetAddrInfo().
There were no remaining locks around __pmGetNameInfo().