QA - mmv pmda - re-install as dso pmda
Fix the overly simple logic in my previous round of commits
... now if the mmv pmda was installed when a QA test starts it
will be installed as a dso pmda after the test ends, independent
of the test passing or failing.

I could be smarter and look at pmcd.agent.type["mmv"] but that
seems overkill until someone comes up with a good case for the mmv
pmda being installed as a daemon pmda ahead of QA running (this
will break expected output for some tests, as Nathan diagnosed,
independent of the logic in this commit).