configure.in - more CFLAGS tweaks
When I move the optimization flags out from $CFLAGS to $CFLAGS_OPT
in the build infrastructure (so I can build/debug code with no
optimization in place), there is a messy cross-dependency between
$CFLAGS use in configure.in and in builddefs.

I missed the fact that -D_FORTIFY_SOURCE=2 _requires_ -O (at least
for some toolchains), and this was causing some malconfiguration
in the cc tests run from configure, specifically the test for the
const (or not) struct dirent parameter to scandir().

This commit fixes that.