Non-root PMDA access control changes
PMDAs may run as root, or the user pcp, or some other user and/or
group (e.g. the dbms PMDAs).

There are two places an arbitrary PMDA needs to be able to write
1) $PCP_LOG_DIR/pmcd to create its log file, and
2) (optionally) $PCP_VAR_DIR/config/pmda to manage any persistant
   instance domain cache files

Both directories need to be mode 1777, but because packaging
restrictions may get in the way, we uncoditionally set the mode
on these two directories in the pmcd start up script.

Some PMDAs may need to take additional care with ownership of
their indom cache files, as the simple PMDA's Install script
demonstrates.