Interfaces for PMDAs making use of client authentication
Pushes the security attributes mechanism into libpcp_pmda.
This will allow authenticated client information, making use
of the existing client context concept, to be used by PMDAs
that want to provide user-specific metric values (for example).
Updated the Linux pmdaproc to use v6 pmda interfaces preparing
it for assimilation down the track.

Added in convenience names to the pmdaInterface structure
so that the version-union struct names can match up with the
PMDA_VERSION macro names as its just alot more intuitive.
Also reworked the semantics of the internal HAVE_V_XXX macros
in libpcp_pmda as they were getting out of hand with addition
of yet another.

Removed an incorrect pmdaMain(3) diagnostics section entry,
and removed references to long-deprecated library routines
(that do not even exist in the code anymore).

Make __pmdaSetupPDU static within libpcp_pmda - it only makes
to use internally.  Modified it to allow manipulation of the
credentials flags to tell pmcd this PMDA is interested in any
authentication information that it sees.

QA test updates to exercise the additional PDU, and deal with
the introduction of a new PMDA interface.  Work remains within
pmcd and libpcp to actually transfer the auth attributes, but
its now largely understood how best to tackle that.