Add support to dbpmda for exercising the attribute callback
New command added into dbpmda to allow testing and debugging
of PMDAs choosing to use the security attributes interface.

Added simplistic support into pmdasample, and added test 712
to exercise the new functionality.

Some code was refactored libpcp to share lookup & reporting
of attribute strings, and attribute:value strings.  This is
used in dbpmda now, as well as in pmdasample and QA sources.