Improve archive integrity checking from libpcp_import
Add new error code PMI_ERR_BADTIMESTAMP for situation where
timestamps presented to pmiWrite() for a single archive are
not monotonic increasing.

Add check in pmiWrite() and Fatal Error message when
PMI_ERR_BADTIMESTAMP is detected.

New qa/527 to exercise this, new output for qa/369.

This addresses some of the issues raised by
https://bugzilla.redhat.com/show_bug.cgi?id=958745