Improvements to the socket-based PMDA mechanisms.
Dave noticed that a PMDA installed as a socket would
sometimes/often fail to complete (accept) a connection
from pmcd, in the pmda-as-server mode (i.e. not started
by pmcd - a little used mode evidentally, but supported
all the same).  Turned out to be the accept(2) being on
a socket that was not necessarily ready - judicious use
of select(2) saves the day, as done in other places.

While there, I noticed the inet and now ipv6 sockets are
listening on any interface, which means they are network-
facing when they really dont need to be.  Make this use
the local loopback interface to tighten this up.  Further,
we only need to listen for a max of 1 connection request -
pmcd - not 5 (cut&paste from elsewhere I imagine).