pmdagfs2: changes based on review feedback on pcp mailing list
Made general improvements with code for the pmda over last time whilst
taking on suggestions made by Nathan Scott. Have improved collection
code for the glock_lock_time trace-point metrics as well as switching
to the provided pmdaCache(3) data structure instead of a home made
linked-list datastructure.

Introduced a new gfs2.control.glock_lock_time metric which can be used
to enable/disable the glock_lock_time statistic collection based on
its value, this can be set using pmStore(3)on this metric. On installs
without the support for trace-point based statistics, the metrics are
disabled by default because the necessary config files will not be found.

Cleaned up the collection loop for the trace-point data, have moved to a
single loop that removes the dependency on allocation of dynamic memory
to achieve collection of the data. These changes will reduce the memory
requirements of the pmda and speeds up the pmda with less iteration of
the data.

Changed the data structure to hold our collected data, we have gone
from home made linked-list to the available pmdaCache(3) this reduces
the amount of code, simplifies reading and reduces again the amount
of iteration of the locks we have (no more remove duplicate and easier
addition through a well defined hash).

Changed the collection rate by changing the calling
gfs2_refresh_lock_time() from once-per each file-system to once for
all file-systems, this reduces the complexity of data storage. Reducing
potential of lost lock data in-between refreshes. Changed the number of
calls to gfs2_refresh_lock_time() from once per each filesystem to now
being called once for all file-systems also removed forgotten testing
code that was left within the gfs2_fetch_refresh function.

Separated instance refresh so that the collection of dev_id is handled
within another function in order to increase code readability. Introduced
the necessary pmStore(3) hooks so that we can use the control based
metrics to handle whether we are collecting statistics from the
trace-point based metrics borrowing from the techniques used in the
linux pmda.

Split the QA files into two halves, the first 654 handles testing of
the base part of the pmda (glstats, sbstats, glocks metrics). 655 then
handles testing for the trace-point based metrics with not_run escapes
in an event that trace-point based metrics are not available for the
install (i.e. Debian unstable?). A new common.gfs2 script handles common
functions used in both of the tests.