Bug fixes and refactoring in gfs2_refresh_lock_time
The cache cull call on the lock_time_indom was being
done inside the loop which accumulates glock stats;
therefore whenever it takes multiple reads to consume
the available glock stats, those from the previous
reads would have been accidentally culled.

One reason this was not immediately obvious, perhaps,
was that the code to extract a single glock stat was
interwoven within gfs2_refresh_lock_time - I've moved
this out into its own function which clarifies things
(to my eyes at least).

It then also became clear that the "hash==NULL" check
here can never fail, since hash is an on-stack buffer,
so I simply removed that (dead) error-checking branch.

Finally, this refactoring also makes the last buglet a
bit clearer - we fopen a file (allocates and returns a
FILE*), then extract the fd via fileno(3), but when we
are cleaning up at the end we call close on the fd and
not fclose on the fp - which leaks the FILE* structure.