Make pmlogger and pmlogextract first class citizens
Some discussion on IRC ended up with the what-a-good-idea
suggestion of making a couple of very important PCP tools
available on the default path.  Previously for historical
reasons (I think) they were squirrelled away in BINADM_DIR.
This was the second time in as many weeks that people were
seeking logextract functionality and didn't know it already
existed cos of its unusual location.

This commit moves the two tools into /usr/bin and installs
back-compat symlinks in the old location for anyone with
hard-coded paths (scripts, etc).

The man pages no longer refer to them in the old locations
anymore as well.  Also, noticed an incorrect reference in
the dbpmda(1) man page to its location being in BINADM_DIR
when in fact it is not - fixed.

This resolves Fedora bug 967709.