Gets the python convenience classes module up and stumbling
Fixes a number of out-of-date parts - particularly around the
handling of pcp.pmapi return values which nowadays use proper
exceptions for the error code.  Lots of c-wrapper issues also
sorted out here though.

Additionally, changes the way the code was pushing timestamps
around to every single group, instead of just keeping this in
one spot in the surrounding group.

Plenty of work remains in this module, however - needs a fair
bit of work to make it pylint clean, testing, etc.