Fixes a sigsegv in pmchart command line error handling
Resolve a problem with the way pmchart iterates over the archive
(and/or host) list from the command line.  Should an error occur
we remove the archive/host but we do not process the remainder of
the list correctly, accidently skipping one element.  The fallout
from that ends up being a sigsegv in the metrics class code later
on (timezone setup, expecting a context to exist which doesn't).

This resolves Fedora bug 967713.