Remove limit in the libpcp import auto-PMID generation scheme
Marko found that no new metrics could be imported into a log
beyond 1024.  Turned out to be a lack of pmiAddMetric checks
for metric counts beyond the PMID "item" space (1<<10).  Its
now making use of the "cluster" space (1<<12) as well, and I
added in a check for the real boundary of auto-generated IDs
which is ~4.2 million metrics.

Test case 537 exercises the fix and makes use of Marko's nice
succinct test script which exposes the problem.

This resolves Fedora bug 968210.