Fix test 528 which attempted to insert two values for one metric
This results in an error (DUP_METRIC_VALUE) when these are in the
same pmResult, exposed by Markos recent improvements to the perl
log import batching.

Also noticed a floating-point rounding issue that would have made
this test fail on some platforms, so changes to just use strings
everywhere, since thats simpler and still exercises all of the
functionality we're after here.