pmdagfs2: gfs2_store() related cleanup and fixes to QA tests qa/654 qa/654
Removed the needless on-stack "value" local variable in gfs2_store() and
all other related code belonging to it.

Fixed both QA tests to exercise the pmda 654 runs and can now be passed on
any machine that has the gfs2 modules through the use of loop-back mounted
filesystems (expects no mounted gfs2 filesystem). 654.out has been updated
in order to match the expected outcome of running this test.

655 has been updated to be able to exercise the gfs2_glock_lock_time metrics.
In order to simulate a correct run, the test machine requires gfs2 modules to
be loaded, an install with support for the additional gfs2 tracepoints and
additionally a filesystem must be loaded in a shared cluster environment
configuration. Because a locally mounted filesystem will not work, additional
checks for cman are made otherwise the test will not run. An expected 655.out
has been produced for this test.

common.gfs2 has been updated to reflect the new changes made to both of the
654 and 655 QA tests.