Convert the multi-threaded tests to use runtime feature detection
The multi-threaded tests were all using the presence of a macro in
the installed pcp/impl.h to determine whether they should run.
This is problematic in that the header may be out of date, or may
not be present at all if the devel package not installed.

Now that we have a libpcp runtime feature test mechanism, use it.