Handle non-determinism with loopback devices in gfs2 qa test
The order in which the loopback paths are presented (readdir)
determines the order in which they will be assigned to instance
IDs in pmdagfs2 - this is non-deterministic (depending on the
whims of the loopback device driver, 1 may appear before 0).
Update common.gfs2 to handle this while still representing the
fact that both instances appear and have (non-error) values.