collectl2pcp fixes and test 544 update for less verbosity
Batch up a series of metric groups for pminfo, and checksum
the results rather than dumping everything in the test.

In doing this, noticed that hinv.ndisk was missing from one
of the archives.  Investigation revealed an off-by-one in the
collectl file header parsing, resuling in that metric being
inadvertantly skipped.

In yet further discussions with Mark we noticed the process
metrics proc.psinfo.cmd metric was not exporting the values
as pmdaproc does (should be basename).  So proc.psinfo.psargs
added in with the original cmd value, as nature intended.