Push uid/gid details slightly further down in libpcp (out of impl.h)
This removes the assumption from the __pmAccAddAccount interface that
a uid and gid can be represented by an integer, which is not portable
it turns out (fortunately, we don't need to make assumption as all of
the low-level authentication/SASL2 code fundamentally using values as
strings, which is just right for our needs here too).

Adds configure.in check for uid_t, gid_t and SID (Win32 Security ID).
While in there, added a sys/socket.h AF_UNIX check for Dave.