Update the unix domain sockets configure checking
It turns out that the current mingw headers (esp from the mingw64
project) define AF_UNIX, so testing for this is problematic.  The
structure sockaddr_un is not defined however, so this makes for a
more appropriate test target.

Update the code that directly reference IS_MINGW and update error
messages accordingly.