Correct the logic in pmcd around enabling user/group access controls
Needed to deal differently with the case where no ACLs are present, and
the case where ACLs are present but no account information is available
on the new clients connection.  Also ensure the initial (wildcard) entry
has a default-restrictive set of permissions, not default-to-open.

Test case 546 added to exercise user/group ACL extensions to the pmcd
configuration file parsing (error handling cases in particular).

Finally, update several cases of direct uid/gid comparisons; need to use
the accessor methods if this is ever to work on non-POSIX platforms.