Update test 635 to ensure netstat reported interfaces match ours
It was noticed that on a host where an interface was marked down,
pminfo and netstat output do not match, because netstat will only
report on these interfaces when the -a option is presented.  We
export /proc/net/dev interfaces without that filtering (client is
able to make use of network.interfaces.up to cull the riff-raff).
Easily resolved by adding -a option to netstat now that the issue
is understood.