Functioning instance domains for the pcp.pmda python module
This is the last (for now) piece of the PMDA python wrapping
puzzle.  The python variant of the simple PMDA now operates
correctly for all of the different PDU variants.  Yippee!

Particularly problematic were the pmdaCache instance domains
as there is a bit of a conflict between pythons instance on
maintaining type information, and the (void*) parameters we
use in the C API so that PMDAs can store whatever they like.
This is solved using ctypes, with the side-effect that only
ctypes values can be used with pmdaCache indoms.

Other issues resolved include: ensure the instance callback
declaration in python PMDAs matches the C calling convention
(must declare the passed-in indom serial), big improvements
in the python stacktrace loggin when exceptions thrown, and
correct the argtypes declarations for a couple of the cache
interface routines from libpcp_pmda (void** vs void*).