Fixes to the python pmda interface code to allow old python versions
Workaround several issues from old python versions (RHEL5, so 2.4.3).
There is no Py_ssize_t there, so conditionally dodge that bullet.  The
buffer interface was reworked in 2.6+ evidentally (or at least, back-
ported there) so any use of PyBUF_SIMPLE and the PyObject_GetBuffer,
PyObject_CheckBuffer and so on must also be guarded.

Finally, PyArg_ParseTuple takes (const char *) in latest version, but
was not const back in the day, so avoid that warning as well.