pmdagfs2: Addition of documentation and manpage entry
Created a manpage entry for the GFS2 PMDA, following the standard template
that is used by other PMDA's. The usual sections are included on how to
install, remove etc. with the addition of a section mentioning the
requirement for a clustered setup for certain metrics to be able to provide
metric data.

Updated the commenting in qa/654 to explain better the expected setup for
the test to be run and to make clear that the test does not query metrics
which require a clustered environment.

Added extra information about for qa/655 with regards to a test setup and
provided information along with a link to fabbione's cbox (cluster in a
box), providing an easy route to setting up the required test environment
though a virtualised two node cluster using KVM. Cbox is available at
https://github.com/fabbione/cbox.

Created README to give an overview of the PMDA, it's requirements, how to
install, remove along with some basic troubleshooting ideas. Have also
touched upon how to go about using pmstore with regards to the
configurable metrics.

Slight tweak in GNUmakefile to guard against the deletion of the README
file when issuing "make clean".