pmwebapi: support more general pcp hostname specs, and HTTP Basic authentication
When a web client uses a hostspec string to create a new context, and
if that string includes PCP_ATTR_USERNAME/PASSWORD information, make
the new webcontext require use of matching HTTP Basic authentication
for it.

Renamed the /pmapi/context?hostname= key to ?hostspec= to emphasize
the generality.  (The hostname= one is supported for backward
compatibility.)