pmdagfs2: gcc -Wmaybe-uninitialized compilation warnings fix
Made changes in lock_time.c to ensure that glockA and glockB are initialized
as zero prior to having values from the hash-map containing glocks for the
given GFS2 filesystem assigned to them. This silences the gcc warnings of
potential uninitialized  variables in the glock comparison function during
compilation of code.