Add in automated checkers for all the past security flaws
Had permission to merge these tests many months ago, but no time
nor pressing need to merge back given the other testing that was
done (esp valgrind-based pducrash and so on).  It is handy to do
now though to sync up with other QA folks here as well as ensure
ongoing coverage in dev branches.