Resolve source layout inconsistency between pmlogger/pmie/pmcd
For some strange reason a number of pmlogger scripts has been
put in a different location (pmlogctl) to the pmcd and pmie code
which has scripts in top level then source in a "src" directory.
This commit makes pmlogger toe the line, keeping everything the
same for ease of locating the various helper scripts/sources.