Reserve a QA test number for pmdagluster checkout
Adds in initial version of a gluster filesystem agent test.
This uses dbpmda and the pmdagluster env override mechanism
to pass test data into the PMDA, allowing it to be tested
in arbitrary configurations even without a gluster setup.

Some initial data for a simple setup included, but expect to
extend over time to some more interesting configurations.