Resolves all known remaining issues in pmdagluster
Much work done in the QA testing space, ensuring that all of the
instance caching works correctly, fetch and store callbacks are
now extensively tested and the data they return verified.  Several
bug fixes in the PMDA required to get to that point and some metric
renumbering (cluster ID split in particular) performed to make some
things simpler in the fetch callback.