pmlogextract - fix minor bug for empty instance domains
If a metric with an instance domain appears in an input archive,
but the corresponding instance domain is empty in the input
archive, then pmlogextract would abort.

This is wrong, as some metrics can have an instance domain that
is currently (or even always) empty.

The fix is to check the result from pmGetInDomArchive() and
if it fails with PM_ERR_INDOM_LOG, then silently continue.
The metrics in the output archive will correctly have no values
in their pmResults for these metrics and the instance domain will
also be empty in the output archive.