Introduce a qa mechanism for switching to/from a writable primary logger
With recently increased restrictions on access to primary loggers (pmlc),
a number of QA tests have begun failing.  This is good in that it shows the
default restrictions are working, but bad in that the original intention of
the tests is no longer being exercised.

This commit adds a common mechanism for temporarily changing the default
logger config to allow local write access via pmlc, and then change it back
(typically at the end of the test in the trap handler).

Initially, test 032 is converted to using it - more to follow soon though.