Fix test program hangs in getopt on some architectures
On ppc64 and s390x the test programs unpack.c and pmdacache.c hang in
the call to getopt(3).  Since getopt is declared with int as its return
value and when there are no more arguments on s390x and ppc64, c = 255
instead of expected c = EOF = -1, therefore while loop never quits.

When c is declared as int, testcases 341 and 404 pass as on x86_64.