Correct the default install.sh invocation after recent changes
Need to default to a root:root install for all files still, this
was inadvertantly changed recently.  We also were inadvertantly
using a no-longer-available macro as the default file owner which
ended up with invalid install lines due to:

    INSTALL = $(INSTALL_SH) -o @pcp_owner@ -g pcp

This now uses the PCP_{USER,GROUP}_INSTALL macros which provide
the expected defaults from the configure.in phase.