pmdagfs2: Improvements for distributions without all gfs2 trace-points
Improved error handling for distributions that do not currently have
support for the full set of GFS2 trace-points.

Additional cases are used within all metrics fetch functions to check
for a pre-determined value which is set if the trace-point does not
exist on the system. This sets the metric values as "No Value(s)
Available" and is a cleaner output than zeros in these cases.

Also a little spring cleaning with some of the coding conventions.

Additional: Fixed issue where direct lookup without bounds check
could result in invalid array position being read.