pmdagfs2: qa improvements to guarding and compatible setups
Have reduced the number of tests run on qa/654 so that it can be run
on most installations with gfs2-utils present. Checked against Fedora
16-19, Ubuntu 13.04, Centos 6.4 and Debian 7 with Debian 7 being the
most restrictive through only having support for gfs2.glock metrics.

Included additional checks in qa/655 to test for the mounting of debugfs
and to attempt to mount it in the event that it is already not. Have
re-added added additional comments on suggestions how to quickly setup a
configuration for testing 655 in qemu-kvm using the cbox application.

Added additional functions to common.gfs2 that test for the availability of
mkfs.gfs2 on the test system (part of gfs2-utils) as well as attempting to
mount debugfs on distributions that do not have it mounted by default.

qa/654.out reflects the changes made to qa/654 in order to verify a correct
run.

Additional: Re-added non-deterministic view of loopN on tests in qa/654 and
added pmstore checking in qa/655 to ensure the functionality is checked in
future (testing gap).