Additional protective measures preventing system config overwrites
This adds a command line option to pmlogconf and pmieconf which is
now always used the the check scripts.  It places magic inside the
generated configuration file which is used to make sure that only
configuration files that were originally generated as part of the
automatic pmlogger/pmie enabling will ever be touched by those two
auto-configure tools.

It is another in a series of defenses now, aimed at preventing any
kind of accidental foot-shooting episodes when the default system
pmlogger/pmie processes fire up.