Ensure we do not force unix domain sockets down the SASL path
Somehow I lost this change on an earlier commit.  It removes an
earlier (incorrect) association I made that unixsock/local attrs
on a connection should force the authentication part of the over-
the-wire protocol exchange to happen.  If we do that, we end up
in SASL-land for unix domain sockets, which we do not want/need.

This fixes test qa/831 once more, which spontaneously combusted
as a result of this change going AWOL.