Rework (and fix) the /proc/net/snmp metric extraction code
As recently discovered by the Red Hat QE folk (running qa/126
on recent kernels) the /proc/net/snmp file has been extended
to add checksum error counters to several message types.  For
most cases this was fine and the code still worked, except in
the case of the Icmp values.  In that case, the new value was
added in the middle of the existing set and not the end.

This has exposed an assumption in the code where it chose to
ignore the header information and assumed all new values would
be appended.  This is no longer feasible, and is tackled here.

We now parse the headers (every other line) and match them up
to the values on the following lines.  The new checksum error
metrics are added and reported as not-available on all kernels
predating this feature.

Implementation-wise, these metrics are handled more like the
vmstat metrics now, so there was some minor cleanup resulting
from a closer look at that code (esp. removed a dup'd global).

Finally, we now had the option of also implementing the icmpmsg
values which previously we could not due to ignoring the header
(these values are more dynamic, and not a fixed set of values).
This also has been done, and a new indom is added to handle the
more dynamic nature of these values (see network.icmpmsg.*).