Fix loop termination check for ICMP stats, older kernels
In the reworked /proc/net/snmp parsing code, as a result
of the incsumerrors field being optional (older kernels),
we need to be careful when matching up header and values
columns against the expected set.  This fixes incorrect
behaviour where the final field in the row was missed, in
the case where there is a missing field earlier on in the
row (as is the case with the ICMP metric values).