Install a default sasl2 configuration file for pmcd
Following the lead of other (more experienced) projects
using SASL2 authentication, such as libvirt, install a
default configuration file for pmcd and SASL.

This configuration (like other projects) enables only
the digest-md5 plugin by default.  We may want to also
allow "plain", perhaps, but for now play it safe.

A companion tutorial for "secure sockets", decribing the
details of this configuration and many other aspects of
authentication in PCP, is being committed in the pcp-doc
package as well.