Rework the interactive prompting code for SASL callbacks
Major rework of the SASL interactive prompting code.  This is
now much simplified in the split between command line options
handling and filling-in-the-missing-bits interactively (incl.
password prompting).  Removes the use of the obsolete getpass
libc function.

Additional SASL mechanisms now appear to work correctly (happy
days!) and the "plain" method has now been used for successful
authentication, along with a sasldb for user accounts.

Added some comments to the default sasl2 configuration file,
particularly along the lines of file permissions on a sasldb
file that pmcd may have to read.