Non-default permissions artifacts and Debian packaging
This commit tries to enforce consistency between the build,
the Debian packages and the PCP run-time.

In general terms, if the directory mode is 775 or the file mode is
664, then we expect the ownership to be pcp:pcp.  There are some
execeptions but these are noted with large comments.  To enforce
this the makefiles need to explicitly set the mode, owner and
group for the non-default artifacts.  The Debian pcp.postinst
script needs to do _exactly_ the same thing as the settings from
the makefile are stripped by dh_fixperms in the build.  And at
run time we sometimes have to create or recreate things on the fly
and the owner, group and mode here have to play by the same rules.

Finally when we smack these files in qa, we need to be sure
to restore the owner, group and mode to match the makefiles.
The new qa/994 checks a lot of this.  The first batch of QA
changes are in the next commit.

If this survives Debian QA stress testing, I'll be making the same
(philosophical) changes for the other packaging options we support,
one by one.