Get qa/832 passing reliably without prompting
Need to take tighter control on the test environment, in
particular the sasl config needs to be setup to ensure a
specific failure mode.

In the process, make the libpcp console output a bit less
free with the end-of-line output, making interactive auth
input/output a little more neat.

Added a note about saslauthd in the pmcd sasl2.conf file.