Fix pmpost for use by the unprivileged pcp account
As pmpost may nowadays be invoked by pmie running as "pcp",
we need to do a better job of creating the noticeboard file
in terms of filesystem permissions.  This change ensures it
is created owned by root, and writable by group "pcp" (iow,
$PCP_GROUP).

Long-gone are the days when pmpost ran setuid root, so the
man page is also updated to remove those references and to
document this new behaviour.