GNUmakefile - change user for pmda cache
As a general rule, if we need non-root things, then I'd favour
setting uid to pcp _and_ gid to pcp and mode to 775 (dirs) or 664 (files).

This would (a) be a consistent policy with exceptions, rather than totally
ad hoc, and (b) more importantly, means we have independent entitlements
that could be use if we need 'em, e.g. in the case where something _has_
to run as some user other than root or pcp, they could be in group pcp
and still do what needs to be done.