qa/366 - smack nfs2 and nfs3 stats
As we make more use of pmlogconf, there will be additional groups
of metrics expanded by default.  When a group used in this test
goes from being not expanded to expanded the test will fail.

Fixed for the nfs2 and nfs3 metrics.

And document in qa/366 how to handle this sort of issue going forward.