Additional MySQL PMDA status metrics suitable for pmie consumption
1. Added duplicate metrics for some of the status values which are in
the form of YES/NO or ON/OFF. These textual binary values can't be
interpreted by tools such as PMIE, so each of these status values now
have numerical friend.

> mysql.status.slave_running     value "ON"
  > mysql.status.slave_running_num     value 1

> mysql.status.compression     value "OFF"
  > mysql.status.compression_num     value 0

> mysql.slave_status.slave_io_running     value "Yes"
  > mysql.slave_status.slave_io_running_num     value 1

> mysql.slave_status.slave_sql_running     value "Yes"
  > mysql.slave_status.slave_sql_running_num     value 1

> mysql.slave_status.master_ssl_allowed     value "No"
> mysql.slave_status.master_ssl_allowed_num     value 0

2. The status variables were being captured using the MySQL command
"show status". This was replaced with "show global status".

Refer this text from http://dev.mysql.com/doc/refman/5.1/en/show-status.html

<quote>
With the GLOBAL modifier, SHOW STATUS displays the status values for all
connections to MySQL. With SESSION, it displays the status values for
the current connection. If no modifier is present, the default is
SESSION. LOCAL is a synonym for SESSION.
</unquote>

3. Updated the perldoc to give some information on the new metrics.