Usability improvement for the New/Edit Chart Apply button
We now maintain significant state within the New/Edit Chart
dialog after pressing "Apply", so that work can continue to
be performed with the current selected metrics and hierarchy
navigation remaining in place.

Additionally, on Apply with a New Chart, it is more likely
that someone will want to make additions to the chart they
are actively creating, than to create more New Charts (from
bitter experience).  For this problem, we now automatically
switch on-the-fly into Editing mode for the freshly created
chart.

This resolves Fedora bug #957669.