Fix pmchart handling of missing metrics in View files with archives
Properly catch errors being passed back from openView() instead of
ignoring 'em (continuing on in a horribly broken state).  In the
process found that the closed-socket handling in the client-side
pmtime connection was doubling-up (and racey) when tearing down the
live/archive socket connections, so fixed that at the same time.

This fixes Fedora bug #981140.