Fix pmchart handling of total lack of any metrics source at startup
When no pmcd process is available, and no archives have been given,
pmchart struggles on bravely.  This gives the user a chance to use
the tool with remote hosts, or archives added at runtime, which is
good and admirable.

However, there is a couple of issues when pmchart is stopped before
rectifying its lack of any metric sources.  It attempts to make use
of a context when it shouldn't (lack of error check in Qmc lib) and
it pops up a pmquery window to tell us all about it.  The latter is
a particular source of bugs (even in other situations), because it
causes pmchart to fail to exit cleanly, while it waits for the user
to interact with the popup.  This can hang logout within a desktop
environment and other nasties.

Fix the missing error check, and force any final errors onto stderr
rather than any exit-blocking popup during pmcharts exit.

This resolves Fedora bug #957007.