Resolve a qa/831 regression with recent authentication fixes
With an unrelated recent change, we inadvertantly began to send
the uid/gid connection information to PMDAs twice for the unix
domain socket case - once on initial connection, and now also
after credentials PDU arrival.  Test 831 dutifully picked this
up and began failing ... perfect.

We only need to tell the PMDAs once, this change makes it so,
and as a result test 831 is passing once more.