Fix problems in the pmchart Samples / Visible History dialog
There are two problems with the interaction between the main
pmchart window and the samples settings dialog - the values
are reset incorrectly to the default values each time the
dialog is popped up, and sometimes the dialog does not popup
at all as a result of delayed dialog creation.  This commit
fixes both of these problems.

This resolves Fedora bug #968825.