Rework earlier change to pmdaproc.sh tmpfiles
Turns out the issue is more subtle than it seemed, and
we do need two separate set of tmp files at different
times.  Ugh.  Deal with that, and ensure cleanup takes
both into account.  Also fix an assumption that a file
(pmcd.body) would be created, where in some cases it
would not be.