Transition Linux kernel /proc/net/snmp metrics to 64 bits
The kernel exports these values as 64 bit these days, whereas
we extract them as 64 bit and at the last minute fib and say
they're 32 bits.  This is a more noticable problem on 64 bit
big endian hosts, where that final transition often truncates
to zero.  As a result, test qa/216 fails there.

Adds in a logrewrite rule and test qa/827 to exercise it all.