Fix missing cleanup of temporary files in test qa/709
Turns out pmcollectl in test qa/709 was the one creating
mysterious temporary files in the qa directory.  Resolve
this by running the test in a tempdir sandbox - but it'd
be good to check if pmcollectl/pmapi is doing all needed
cleanup for archive folios.  Will seek help from Stan on
that one.