Provide low-risk resolution to the unix:/local: hostname problem
Use the same paths that are used when "localhost" is presented
to pmlogger, to get a meaningful hostname suitable for the log
label (and later use/reporting by client tools).

This avoids the pmcd socket path being written into the log label
as host name.  There is a wider problem here than just the unix:/
local: naming though, lets revisit that in a future release.