pmcd->pmlogger hostname passing
pmlogger prefers to put a relatively unique hostname into the log
header to identify the source of the data, but use of the PMAPI
connection parameters is only a heuristic substitute, as is the
cleverness involving mapping "localhost" to `gethostname`.  This
becomes worse when pmlogger runs with -h local://, where the number of
patterns to match to the heuristic gets silly.  It gets even worse if
TCP port-forwarding is in use, where the connection hostname has
nothing to do with the actual PMCD hostname.

This patch adds a pmcd.hostname metric to the pmcd, which is queried
by pmlogger just as pmcd.timezone.  If present, the pmlogger's
heuristic logfile-header hostname will be replaced with the pmcd's
self-reported one.