rpm packaging and perl module package issues
This change allows us to take finer control of the Requires dependencies
that are auto-generated by rpm ... in the initial case, we dodge the
Spreadsheet::Read dependency for import-sheet2pcp.

In the process, this allows RedHat BZ 830923 and BZ 754678 to be
resolved (once the fedora.spec file is restored to its former glory).