Split the XFS and XFS quota metrics into a separate PMDA
We recently discovered that the XFS quota.* metrics require
root access to function, and thus were inadvertantly broken
by the changes to run pmcd as an unprivileged user.

The xfs.control.reset functionality, for resetting the XFS
kernel statistics back to zero also requires root access.

This change moves all of the XFS metrics out of the Linux
kernel PMDA into a new pmdaxfs, which runs as root.  A log
migration rewriter configuration is also provided for auto
transitioning daily archives.

New QA tests will be added in a followup commit exercising
both the fix to the original problems, and the log rewrite
configuration mechanism.